+++ StabilityMatrix.Core/Models/Update/UpdatePlatforms.cs
@@ -11,6 +11,9 @@ public record UpdatePlatforms
     [JsonPropertyName("linux-x64")]
     public UpdateInfo? LinuxX64 { get; init; }
 
+    [JsonPropertyName("macos-arm64")]
+    public UpdateInfo? MacOsArm64 { get; init; }
+
     public UpdateInfo? GetInfoForCurrentPlatform()
     {
         if (Compat.IsWindows)
@@ -21,6 +24,11 @@ public record UpdatePlatforms
         if (Compat.IsLinux)
         {
             return LinuxX64;
+        }
+
+        if (Compat.IsMacOS && Compat.IsArm)
+        {
+            return MacOsArm64;
         }
 
         return null;